+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release Notes for Craft Generator
 
 ## Unreleased
+- Added the behavior generator. ([#27](https://github.com/craftcms/generator/pull/27))
 - Generator now prompts for new modules’ names. ([#28](https://github.com/craftcms/generator/pull/28))
 - Added the `--with-strict-types` option. ([#29](https://github.com/craftcms/generator/pull/29))
 - Fixed issues with the default `getTriggerHtml()` method output for generated element actions. ([#30](https://github.com/craftcms/generator/pull/30))